requirements.txt
websocket-client>=1.6.0
fastapi>=0.115.0
uvicorn>=0.30.0
redis>=5.0.0
chromadb>=0.5.0
